Don't check Vec length, but instead convert Vec to &[&str; 4]
This means that rather than checking an invariant but still continuing
to use the type that does not guarantee that invariant, we extract an
array that's guaranteed to be length 4 at compile time. This means that
the invariant becomes intrinsic in the type itself, rather than in code
earlier in the function.