Serve cert challenge from http
For some reason after a recent update (or, rather, an update in the last
two months ten days) certbot will no longer renew certificates if the
challenge is served under https. This is especially weird, since it
doesn't explain that this isn't allowed, and I can access the challenges
manually using curl/the browser.

Either way, it's probably best to serve the certificate renewal challege
without needing the current cert to be functioning.